job_filter_app/requirements.txt
# Core dependencies for job scraping
pandas>=1.5.0
pyyaml>=6.0
requests>=2.28.0

# JobSpy dependencies (already included in parent directory)
# No additional packages needed for basic scraping functionality
